requirements.txt
fastapi>=0.115.12
uvicorn>=0.34.2
motor>=3.7.1
pydantic>=2.11.4
email-validator>=2.2.0
fastapi-mcp==0.3.4
passlib==1.7.4
pymongo==4.13.0
python-dotenv==1.1.0
python-jose==3.4.0